requirements.txt
streamlit
fastapi
uvicorn
openai
langchain
chromadb
python-dotenv
pydantic

